+++ core/core-auth/src/main/java/com/barlow/core/auth/authentication/access/AccessTokenPayload.java
@@ -4,4 +4,8 @@ public record AccessTokenPayload(
         Long memberNo,
         String role
 ) {
+
+    public static AccessTokenPayload ofGuest(Long memberNo) {
+        return new AccessTokenPayload(memberNo, "GUEST");
+    }
 }
