+++ core/core-auth/src/main/java/com/barlow/core/auth/authentication/access/MemberPrincipal.java
@@ -1,4 +1,6 @@
-package com.barlow.core.auth.principal;
+package com.barlow.core.auth.authentication.access;
+
+import com.barlow.core.auth.authentication.core.Principal;
 
 public class MemberPrincipal extends Principal {
