+++ core/core-auth/src/main/java/com/barlow/core/auth/authentication/core/AuthenticationExceptionType.java
@@ -4,6 +4,5 @@ public enum AuthenticationExceptionType {
 
     INVALID_CREDENTIAL,
     EXPIRED_CREDENTIAL,
-    PERMISSION_DENIED,
     ;
 }
