+++ core/core-auth/src/main/java/com/barlow/core/auth/authentication/core/Authenticator.java
@@ -1,7 +1,5 @@
 package com.barlow.core.auth.authentication.core;
 
-import com.barlow.core.auth.principal.Principal;
-
 public interface Authenticator<C extends Credential, P extends Principal> {
 
     P authenticate(C credential);
