--- core/core-auth/src/main/java/com/barlow/core/auth/authorization/AuthorizableMemberInfo.java
@@ -1,7 +1,0 @@
-package com.barlow.core.auth.authorization;
-
-public record AuthorizableMemberInfo(
-        String memberNo,
-        String role
-) {
-}
